Enhanced Gravity Strike with comprehensive debugging and direct asset loading framework
- Fixed inverted conditional logic that caused program to fall through
- Added verbose flag (-v) for basic debugging output
- Added extra verbose flag (-v2) for detailed data structure dumping
- Enhanced mouse initialization to handle remote/headless environments
- Implemented DIRECT_ASSET_LOADING compile flag framework
- Added comprehensive asset reverse engineering documentation
- Improved data file loading diagnostics with detailed error reporting
- Maintained backward compatibility with existing codebase

This commit adds extensive debugging capabilities to help identify and resolve
the data file loading issues that prevent the game from running properly.

+++ src/gsnetw.cpp
@@ -426,7 +426,7 @@ void gsNetwork::create_shipfile_packet(BitStream *p, PlayerID player_id)
     {
       packfile_password(globals->pwd);
       PACKFILE *shp = pack_fopen(shippath, "rp");
-      pack_fread(shipbuf, sizeof(char), shipsize, shp);
+      pack_fread(shipbuf, sizeof(char), (int)shipsize, shp);
       pack_fclose(shp);
       packfile_password(NULL);
     }
